fix(): Bug Dashboard + Git
Bug parseDate de Dashboard
Add all the files to git

+++ src/main/java/com/excilys/service/ComputerService.java
@@ -1,36 +1,61 @@
 package com.excilys.service;
 
-import java.sql.SQLException;
 import java.util.List;
 
-import com.excilys.dao.ComputerDAO;
+import com.excilys.exception.ServiceException;
 import com.excilys.model.Computer;
+import com.excilys.persistance.dao.ComputerDAO;
 
-public class ComputerService {
+public enum ComputerService implements IService<Computer, Long>{
+	INSTANCE;
 	
 	private ComputerDAO computerDAO;
 	
-	public ComputerService() {
-		computerDAO = new ComputerDAO();
+	private ComputerService() {
+		computerDAO = ComputerDAO.INSTANCE;
 	}
 	
-	public Computer getbyId( Long id) throws SQLException {
-		return computerDAO.getbyId(id);
-	}
+//	@Override
+//	public Computer getbyId( Long id) {
+//		if (id <= 0) {
+//			throw new ServiceException("ComputerService_getbyId - Invalid id!");
+//		}
+//		return computerDAO.getbyId(id);
+//	}
 	
-	public List<Computer> getAll() throws SQLException {
+	@Override
+	public List<Computer> getAll() {
 		return computerDAO.getAll();
 	}
 	
-	public Long create( Computer entity) throws SQLException {
-		return computerDAO.create(entity);
+	@Override
+	public void create( Computer entity) {
+		if (entity == null) {
+			throw new ServiceException("ComputerService_create -  No entity!");
+		}
+		computerDAO.create(entity);
 	}
 	
-	public void update( Computer entity) throws SQLException {
+	@Override
+	public void update( Computer entity) {
+		if (entity == null) {
+			throw new ServiceException("ComputerService_update -  No entity!");
+		}
 		computerDAO.update(entity);
 	}
 	
-	public void delete( Long id) throws SQLException {
+	@Override
+	public void delete( Long id) {
+		if (id <= 0) {
+			throw new ServiceException("ComputerService_delete - Invalid id!");
+		}
 		computerDAO.delete(id);
+	}
+
+	public Computer getById(Long id) {
+		if (id <= 0) {
+			throw new ServiceException("ComputerService_getbyId - Invalid id!");
+		}
+		return computerDAO.getbyId(id);
 	}
 }
